% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.treats.R
\name{ace.treats}
\alias{ace.treats}
\title{ace.treats}
\usage{
ace.treats(
  treats,
  traits,
  root.est = "debug",
  param.est = "debug",
  ...,
  replicates = 1
)
}
\arguments{
\item{treats}{a \code{treats} object with traits and a tree (you can use \code{\link{make.treats}} to format your data).}

\item{traits}{a \code{traits} object to simulate the trait (see \code{\link{make.traits}}).}

\item{root.est}{can be either the name of the algorithm to estimate the root value(s), a predetermined set of root values, or a personalised function to estimate them (see details).}

\item{param.est}{can be either the name of the algorithm to estimate traits parameters, a predetermined set of trait parameters, or a personalised function to estimate them (see details).}

\item{...}{any additional arguments to pass to fun.}

\item{replicates}{the number of replicates to diagnose the ancestral states estimations}
}
\value{
This function outputs a treats object with the tree (\code{$tree} a list of traits estimates (\code{$data})) as well as a diagnosis of the estimations (\code{$diagnosis}).
}
\description{
Ancestral characters estimations on treats
}
\details{
This function
}
\seealso{
\code{\link{make.traits}} \code{\link{make.treats}} \code{\link{map.traits}} \code{\link{treats}}
}
\author{
Thomas Guillerme
}
