% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dads.R
\name{dads}
\alias{dads}
\title{Diversity and disparity simulator}
\usage{
dads(
  stop.rule,
  bd.params,
  traits = NULL,
  modifiers = NULL,
  events = NULL,
  null.error = FALSE
)
}
\arguments{
\item{stop.rule}{The rules on when to stop the simulation (see details).}

\item{bd.params}{A \code{"dads"} \code{"bd.params"} object or a named list of parameters for the birth death process (see details).}

\item{traits}{A \code{"dads"} \code{"traits"} object (see \code{\link{make.traits}}).}

\item{modifiers}{A \code{"dads"} \code{"modifiers"} object (see \code{\link{make.modifiers}}).}

\item{events}{A \code{"dads"} \code{"events"} object (see \code{\link{make.events}}).}

\item{null.error}{Logical, whether to return an error when the birth death parameters fails to build a tree (\code{FALSE}; default and highly recommended) or whether to return \code{NULL} (\code{TRUE}). Can also be set to a integer value for the numbers of trials (see details).}
}
\description{
Simulating diversity and trait disparity
}
\details{
\code{stop.rule}: The rule(s) for when to stop the simulation. When multiple rules are given, the simulation stops when any rule is broken. The allowed rules are:
\itemize{
  \item \code{max.taxa}   The maximum number of taxa to reach (including extinct ones).
  \item \code{max.living} The maximum number of living (i.e. non extinct) to reach.
  \item \code{max.time}   The maximum amount of phylogenetic time to reach.
}

\code{bd.params}: This can be either a \code{"dads"} \code{"bd.params"} object (see \code{\link{make.bd.params}}) or a list of named parameters. The allowed parameters are:
\itemize{
  \item \code{speciation} The speciation parameter value.
  \item \code{extinction} The extinction parameter value.
}
By default, this parameter is set to \code{bd.params = list(speciation = 1)}

If \code{null.error} is set to a numeric value, the function will run multiple times until a correct tree is generated. Using this option can greatly increase computational time!
}
\examples{
## Setting some pure birth tree (no extinction) parameters
my_bd_params <- list(speciation = 1)
## Setting some stopping rule (stop when reaching 10 taxa)
my_stop_rule <- list(max.taxa = 10) 

## Run a birth tree without traits
a_tree <- dads(bd.params = my_bd_params,
               stop.rule = my_stop_rule)
## Plotting the results
plot(a_tree)

## Adding some extinction parameter
my_bd_params$extinction <- 1/3

## Getting a simple trait (default Brownian motion)
my_trait <- make.traits()

## Run a birth-death tree with traits
dads(bd.params = my_bd_params,
     stop.rule = my_stop_rule,
     traits    = my_trait)

## TODO: more examples


}
\author{
Thomas Guillerme
}
