% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRitreats.R
\name{dispRitreats}
\alias{dispRitreats}
\title{dispRity to treats}
\usage{
dispRitreats(data, scale.tree = TRUE)
}
\arguments{
\item{data}{an output from \code{treats} containing tree and traits data.}

\item{scale.tree}{logical, whether to scale the trees between 1 and 0 (\code{TRUE}; default) or not (\code{FALSE})}
}
\description{
Converts \code{treats} data output to be passed to \code{\link[dispRity]{dispRity}}
}
\details{
The tree and data parts can be passed to the \code{dispRity} package functions \link[dispRity]{dispRity}, \code{\link[dispRity]{chrono.subsets}} or \link[dispRity]{custom.subsets} as the \code{data} argument (if needed, the \code{tree} argument is automatically recycled).
The \code{scale.tree} function is highly recommended if trees have various root ages.
}
\seealso{
\code{\link{treats}} \code{\link[dispRity]{dispRity}} \code{\link[dispRity]{chrono.subsets}} \code{\link[dispRity]{custom.subsets}}
}
\author{
Thomas Guillerme
}
