% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.fossil.dads.R
\name{drop.fossil.dads}
\alias{drop.fossil.dads}
\title{Drop fossils}
\usage{
drop.fossil.dads(phy, tol = 1e-08)
}
\arguments{
\item{phy}{\code{dads} data.}

\item{tol}{a numeric value giving the tolerance to consider a species as extinct.}
}
\description{
Remove fossils from \code{dads} objects or \code{phylo} objects (using \code{\link[ape]{drop.fossil}}).
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
