% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.things.R
\name{drop.things}
\alias{drop.things}
\alias{drop.singles}
\alias{drop.fossils}
\alias{drop.livings}
\title{Drop things from a tree}
\usage{
drop.singles(dads)
}
\arguments{
\item{dads}{\code{dads} data.}

\item{what}{what to drop. Can be \code{"fossils"}, \code{"livings"} or \code{"singles"} (non-bifurcating nodes).}
}
\description{
Remove fossils or living species or non-bifurcating nodes (singles) from \code{dads} objects or \code{phylo} objects.
}
\details{
\emph{NOTE} that dropping living or fossils species DOES NOT drop associated internal nodes and edge lengths. To drop both fossil/living taxa AND internal nodes, you can use for example: \code{drop.things(drop.things(my_data, what = "fossils"), what = "singles")}.
}
\author{
Thomas Guillerme
}
