% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.conditions.R
\name{events.conditions}
\alias{events.conditions}
\alias{condition}
\alias{condition,}
\alias{time.condition,}
\alias{taxa.condition,}
\alias{trait.condition}
\title{events.conditions}
\usage{
condition(x, condition, ...)
}
\arguments{
\item{x}{the variable to reach for satisfying a condition (see details)}

\item{condition}{the logical function for triggering the condition (e.g. `<`, `==`, `!>`, etc...).}

\item{...}{any optional argument specific for that condition (see details)}
}
\description{
Inbuilt conditions functions for helping designing events
}
\details{
The following functions allow to design specific conditions for events:

\itemize{
     \item{\code{time.condition}}: a conditional function based on the time \code{x}. Typically this can be translated into "when time reaches the value x, trigger a condition" (see \code{\link{make.events}}). There is no optional argument for the function.

     \item{\code{taxa.condition}}: a conditional function based on the number of taxa \code{x}. Typically this can be  translated into "when the number of taxa reaches the value x, trigger a condition" (see \code{\link{make.events}}). This function has one optional argument:
     \itemize{
         \item{living}, a \code{logical} argument whether to consider the number of taxa alive when the condition is checked (default: \code{living = TRUE}) or whether to consider all the taxa simulated so far (\code{living = FALSe}).
     }

     \item{\code{trait.condition}}: a conditional function based on the value \code{x} of one or more traits. Typically this can be  translated into "when a trait reaches a value x, trigger a condition" (see \code{\link{make.events}}). This function has three optional argument:
     \itemize{
         \item{trait}, one or more \code{integer} or \code{numeric} value designating the trait(s) to consider. By default, \code{trait = 1}, thus considering only the first trait to trigger the condition.
         \item{what}, a \code{function} designating what to select from the trait values. By default \code{what = max} to select the maximal value of the trait when the condition is triggered (but you can use any function like \code{\link[base]{min}}, \code{\link[base]{mean}}, \code{\link[stats]{sd}}, etc. or provide your own function).
         \item{absolute}, a \code{logical} designating to consider absolute trait values (\code{TRUE}) or not (default; \code{FALSE}).
     }
}

More details about the \code{events} functions is explained in the \code{dads} manual: \url{http://tguillerme.github.io/dads}.
}
\examples{

}
\seealso{
\code{make.events}
}
\author{
Thomas Guillerme
}
