% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{events}
\alias{events}
\alias{mass.extinction}
\alias{founding.event}
\title{events}
\usage{
mass.extinction(what, when, how, add, test)
}
\arguments{
\item{what}{}
}
\description{
Different events for the birth death process implemented in dads.
}
\details{
More details about the \code{modifiers} functions is explained in the \code{dads} manual: \url{http://tguillerme.github.io/dads}.
}
