% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.dads.R
\name{make.dads}
\alias{make.dads}
\title{Make a dads object}
\usage{
make.dads(tree, data)
}
\arguments{
\item{tree}{a phylogenetic tree.}

\item{data}{a dataset of traits, either a \code{matrix} with column names or a named \code{vector}.}
}
\description{
Combines a tree and some associated data into a dads object (e.g. for plotting)
}
\examples{
## Creating a random tree
my_tree <- rtree(5)
## Adding node labels
my_tree$node.label <- letters[1:4]
## Creating a random dataset
my_data <- matrix(rnorm(9),
    dimnames = list(c(my_tree$tip.label, my_tree$node.label)))
## Creating the dads object
my_dads <- make.dads(tree = my_tree, data = my_data)
plot(my_dads)

}
\seealso{
\code{\link{dads}} \code{\link{plot.dads}}
}
\author{
Thomas Guillerme
}
