% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.events.R
\name{make.events}
\alias{make.events}
\title{make.events}
\usage{
make.events(
  target,
  condition,
  modification,
  add,
  test = TRUE,
  replications = 0
)
}
\arguments{
\item{target}{What to modify, can be \code{"taxa"}, \code{"bd.params"}, \code{"traits"} or \code{"modifiers"} (see details).}

\item{condition}{A function returning a logical to trigger the event (see details).}

\item{modification}{The modification to bring through the event (see details).}

\item{add}{Whether to add this modifier to an \code{"events"} object.}

\item{test}{Logical whether to test if the events object will work (default is \code{TRUE})}

\item{replications}{A numeric or integer value for repeating the event (by default, the event is not repeated: \code{repeat = 0}).}
}
\description{
Making events objects for dads
}
\details{
\code{target} is a \code{character} to designate what will be affected by the event. It can be either \code{"taxa"}, \code{"bd.params"}, \code{"traits"} or \code{"modifiers"}.

\code{condition} must be a function that returns a \code{logical} value and intakes any of the following arguments: \code{bd.params}, \code{lineage}, \code{traits} and \code{time}.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
