% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.modifiers.R
\name{make.modifiers}
\alias{make.modifiers}
\title{make.modifiers}
\usage{
make.modifiers(when, from, to)
}
\arguments{
\item{when}{When to apply the modifier}

\item{from}{Optional, from which value to modify it (if left missing the current value is used)}

\item{what}{What value to modify}

\item{how}{To which value}

\item{modifier.args}{An optional named list of optional arguments for the trait process}

\item{names}{The optional names of the modifiers (generated by default from the process)}

\item{test}{Logical whether to test if the modifiers object will work (default is TRUE)}
}
\description{
Making modifiers objects for dads
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
