% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.traits.R
\name{map.traits}
\alias{map.traits}
\title{Maps a trait on a tree}
\usage{
map.traits(traits, tree)
}
\arguments{
\item{traits}{A \code{"traits"} object (see \code{\link{make.traits}}).}

\item{tree}{A \code{"phylo"} or \code{"multiPhylo"} object.}
}
\value{
A \code{"treats"} object containing the tree and the traits.
}
\description{
Simulates one or more trait specified through a "traits" onto one or multiple trees.
}
\details{
This function simulates the trait(s) on the tree using the tree's branch length.
}
