% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{modifiers}
\alias{modifiers}
\alias{branch.length}
\alias{selection}
\alias{speciation}
\alias{branch.length.trait}
\alias{speciation.trait}
\title{Modifiers}
\usage{
branch.length(bd.params = NULL, lineage = NULL, trait.values = NULL, modify.fun = NULL)

selection(bd.params = NULL, lineage = NULL, trait.values = NULL, modify.fun = NULL)

speciation(bd.params = NULL, lineage = NULL, trait.values = NULL, modify.fun = NULL)
}
\arguments{
\item{bd.params}{A named list of birth death parameters (see details).}

\item{lineage}{A named list containing the lineage data (see details).}

\item{trait.values}{A matrix containing the trait values (see details).}

\item{modify.fun}{A list of internals functions that can modified by \code{events} (see details).}
}
\description{
Different modifiers for the birth death process implemented in dads.
}
\details{
\code{bd.params} can be either a named list of parameter values (e.g. \code{list("extinction" = 0, "speciation" = 1)}) but it is typically handled internally from a \code{"dads"} \code{"bd.params"} object.

\code{modifiers} are functions passed to the birth death process in \code{\link{dads}} to either generate the branch length (named \code{branch.length} and similar) or to decide whether to speciate or go extinct (named \code{speciation} and similar).

For user defined functions, the \code{modifiers} \emph{must} have at least the arguments described above. For safety, we suggest setting these arguments to \code{NULL}.

The pre-build \code{modifiers} in the \code{dads} package are (so far):

\itemize{

     \item \code{branch.length} the simple branch length generator that randomly gets a numeric value drawn from the exponential distribution (\code{\link[stats]{rexp}}) with a rate equal to the number of taxa (\code{lineage$n * bd.params$speciation + bd.params$extinction}).

     \item \code{branch.length.trait} a modification of the \code{branch.length} \code{modifier} where the resulting branch length is changed by \code{modify.fun$modify} if the parent trait(s) meet the condition \code{modify.fun$condition}.

     \item \code{selection} a function returning a randomly sampled integer among the number of taxa available.

     \item \code{speciation} a function returning \code{TRUE} (speciation) if a random uniform number (\code{\link[stats]{runif}}) is smaller than the ratio of speciation by speciation and extinction (\code{bd.params$speciation / (bd.params$speciation) + bd.params$extinction}). If it's bigger, the function returns \code{FALSE} (exinction).

     \item \code{speciation.trait} a modification of the \code{speciation} \code{modifier} where the random uniform number is changed by \code{modify.fun$modify} if the parent trait(s) meet the condition \code{modify.fun$condition}.

 }

More details about the \code{modifiers} functions is explained in the \code{dads} manual: \url{http://tguillerme.github.io/dads}.
}
