% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dads.R
\name{plot.dads}
\alias{plot.dads}
\title{plot}
\usage{
\method{plot}{dads}(x, trait = 1, plot.tree = TRUE, simulations = 50)
}
\arguments{
\item{x}{\code{dads} data.}

\item{trait}{which trait to plot (default is \code{1}).}

\item{simulations}{if the input is a \code{dads} \code{traits} object, how many replicates to run (default is \code{100}).}

\item{...}{}
}
\description{
plot
}
\examples{
plot.dads()

}
\seealso{

}
\author{
Thomas Guillerme
}
