% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simulation.R
\name{plot.simulation}
\alias{plot.simulation}
\title{Plot simulation}
\usage{
\method{plot}{simulation}(data, cent.tend = mean, quantiles = c(95, 50), ...)
}
\arguments{
\item{data}{the simulated matrix}

\item{cent.tend}{the central tendency (default is \code{\link[stats]{mean}})}

\item{quantiles}{the quantiles to display (default is \code{c(95, 50)})}

\item{...}{any additional argument to be passed to \code{\link[graphics]{plot}}.}
}
\description{

}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
