% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dads.R
\name{print.dads}
\alias{print.dads}
\title{Prints a \code{dads} object.}
\usage{
\method{print}{dads}(x, all = FALSE, ...)
}
\arguments{
\item{x}{A \code{dads} object.}

\item{all}{\code{logical}; whether to display the entire object (\code{TRUE}) or just summarise its contents (\code{FALSE} - default).}

\item{...}{further arguments to be passed to \code{print} or to \code{print.dads}.}
}
\description{
Summarises the content of a \code{dads} object.
}
\seealso{
\code{\link{dads}}
}
\author{
Thomas Guillerme
}
