% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.fun.R
\name{sample.fun}
\alias{sample.fun}
\title{Sample function}
\usage{
sample.fun(fun = rnorm, ..., joint = FALSE)
}
\arguments{
\item{fun}{A function from which to sample (default is \code{\link[stats]{rnorm}}).}

\item{...}{Any additional arguments to be passed to fun}

\item{joint}{Logical, whether to estimate both birth and death parameter jointly with speciation > extinction.}
}
\description{
Samples a single numeric value from a function
}
\examples{
## Sampling a single value from a normal distribution
sample.fun()

## Plotting a lognormal distribution
plot(density(replicate(100, sample.fun(rlnorm))))

}
\seealso{

}
\author{
Thomas Guillerme
}
