% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{snapshot}
\alias{snapshot}
\title{Get a snapshot of the tree}
\usage{
snapshot(
  current.time,
  first.time = 0,
  lineage,
  edge.lengths,
  trait.values,
  traits
)
}
\arguments{
\item{lineage}{The internal lineage data list}

\item{edge.lengths}{The internal edge.lengths tracker}

\item{trait.values}{Optional, the internal trait values table}

\item{traits}{Optional, the mechanism to generate traits}

\item{time}{The time of the snapshot}
}
\description{
Creates (and saves) a snapshot of the tree at a particular time point
}
\details{
This function is designed to be used internally in \code{dads} to help \code{modifiers}, \code{traits} or \code{events} objects to get a snapshot of the state of the tree at a required time point.
}
\author{
Thomas Guillerme
}
