% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dads.R
\name{dads}
\alias{dads}
\title{Diversity and disparity simulator}
\usage{
dads(
  stop.rule,
  bd.params,
  traits = NULL,
  modifiers = NULL,
  events = NULL,
  save.steps = NULL,
  null.error = FALSE
)
}
\arguments{
\item{stop.rule}{The rules on when to stop the simulation (see details).}

\item{bd.params}{A \code{"dads"} \code{"bd.params"} object or a named list of parameters for the birth death process (see details).}

\item{traits}{A \code{"dads"} \code{"traits"} object (see \code{\link{make.traits}}).}

\item{modifiers}{A \code{"dads"} \code{"modifiers"} object (see \code{\link{make.modifiers}}).}

\item{events}{A \code{"dads"} \code{"events"} object (see \code{\link{make.events}}).}

\item{save.steps}{Optional, \code{"numeric"} value to save the simulations at specific internal points (this can slow down the algorithm significantly for large trees).}

\item{null.error}{Logical, whether to return an error when the birth death parameters fails to build a tree (\code{FALSE}; default and highly recommended) or whether to return \code{NULL} (\code{TRUE}). Can also be set to a integer value for the numbers of trials (see details).}
}
\description{
Simulating diversity and trait disparity
}
\details{
\code{stop.rule}: The rule(s) for when to stop the simulation. When multiple rules are given, the simulation stops when any rule is broken. The allowed rules are:
\itemize{
  \item \code{max.taxa}   The maximum number of taxa to reach (including extinct ones).
  \item \code{max.living} The maximum number of living (i.e. non extinct) to reach.
  \item \code{max.time}   The maximum amount of phylogenetic time to reach.
}

\code{bd.params}: This can be either a \code{"dads"} \code{"bd.params"} object (see \code{\link{make.bd.params}}) or a list of named parameters. The allowed parameters are:
\itemize{
  \item \code{speciation} The speciation parameter value.
  \item \code{extinction} The extinction parameter value.
}
By default, this parameter is set to \code{bd.params = list(speciation = 1)}

If \code{null.error} is set to a numeric value, the function will run multiple times until a correct tree is generated. Using this option can greatly increase computational time!
}
\examples{
## Setting some pure birth tree (no extinction) parameters
my_bd_params <- list(speciation = 1)
## Setting some stopping rule (stop when reaching 10 taxa)
my_stop_rule <- list(max.taxa = 10) 

## Run a birth tree without traits
a_tree <- dads(bd.params = my_bd_params,
               stop.rule = my_stop_rule)
## Plotting the results
plot(a_tree)

## Adding some extinction parameter
my_bd_params$extinction <- 1/3

## Getting a simple trait (default Brownian motion)
my_trait <- make.traits()

## Run a birth-death tree with traits
dads(bd.params = my_bd_params,
     stop.rule = my_stop_rule,
     traits    = my_trait)

## Simulating a tree using modifiers
## Making a modifier to make speciation trait dependent
my_modifiers <- make.modifiers(branch.length = branch.length.trait,
                               selection     = selection,
                               speciation    = speciation.trait)

## Simulating the tree
dads(stop.rule = list(max.taxa = 20),
     traits = make.traits(),
     modifiers = my_modifiers)

## Run a birth death tree with an event
## 80\% mass extinction at time 4
mass_extinction <- make.events(
                      target       = "taxa",
                      condition    = time.condition(4),
                      modification = random.extinction(0.8))

## Set the simulation parameters
stop.rule <- list(max.time = 5)
bd.params <- list(extinction = 0, speciation = 1)

## Run the simulations
set.seed(123)
results <- dads(bd.params = bd.params,
                stop.rule = stop.rule,
                events    = mass_extinction)
## Plot the results
plot(results, show.tip.label = FALSE)
axisPhylo()


}
\seealso{
\code{\link{plot.dads}} \code{\link{make.traits}} \code{\link{make.modifiers}} \code{\link{make.events}}
}
\author{
Thomas Guillerme
}
