% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.fossil.dads.R
\name{drop.things}
\alias{drop.things}
\alias{drop.fossils}
\alias{drop.livings}
\alias{drop.singletons}
\title{Drop things from a tree}
\usage{
drop.things(dads, what, tol = 1e-08)
}
\arguments{
\item{dads}{\code{dads} data.}

\item{what}{what to drop. Can be \code{"fossils"}, \code{"livings"} or \code{"singletons"} (non-bifurcating nodes).}

\item{tol}{a numeric value giving the tolerance to consider a species as extinct.}
}
\description{
Remove fossils or living species or non-bifurcating nodes (singletons) from \code{dads} objects or \code{phylo} objects.
}
\author{
Thomas Guillerme
}
