% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.events.R
\name{make.events}
\alias{make.events}
\title{make.events}
\usage{
make.events(
  target,
  condition,
  modification,
  add,
  test = TRUE,
  event.name,
  replications = 0,
  additional.args
)
}
\arguments{
\item{target}{What to modify, can be \code{"taxa"}, \code{"bd.params"}, \code{"traits"} or \code{"modifiers"} (see details).}

\item{condition}{A \code{function} returning a logical to trigger the event (see details).}

\item{modification}{A \code{function} bringing the modification to the event (see details).}

\item{add}{Another \code{"events"} to object to add this event.}

\item{test}{A \code{logical}, whether to test if the events object will work (default is \code{TRUE})}

\item{event.name}{Optional, a \code{"character"} string to name the event.}

\item{replications}{A numeric or integer value for repeating the event (by default, the event is not repeated: \code{replications = 0}).}

\item{additional.args}{Optional, a named \code{list} of additional arguments to be used in the event.}
}
\description{
Making events objects for dads
}
\details{
\code{target} is a \code{character} to designate what will be affected by the event. It can be either \code{"taxa"}, \code{"bd.params"}, \code{"traits"} or \code{"modifiers"}.

\code{condition} must be a function that returns a \code{logical} value and intakes any of the following arguments: \code{bd.params}, \code{lineage}, \code{traits} and \code{time}.
}
\author{
Thomas Guillerme
}
