% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.modifiers.R
\name{make.modifiers}
\alias{make.modifiers}
\title{make.modifiers}
\usage{
make.modifiers(
  what,
  when,
  how,
  speciating = NULL,
  waiting = NULL,
  add,
  test = TRUE
)
}
\arguments{
\item{what}{What to modify: can be either \code{"waiting"} or \code{"speciating"}.}

\item{when}{A logical function of how to apply the modification (see details).}

\item{how}{A function of what to return when the condition is met (see details).}

\item{speciating}{A function for the speciating events (ignores all other arguments but \code{add} and \code{test}).}

\item{waiting}{A function for the waiting events (ignores all other arguments but \code{add} and \code{test})..}

\item{add}{Whether to this modifier to a \code{"dads"} \code{"modifier"} object}

\item{test}{Logical whether to test if the modifiers object will work (default is TRUE)}
}
\description{
Making modifiers objects for dads based on an ancestor's (parent) trait.
}
\details{
\code{when} must be a function that inputs at least the parent trait value(s) as \code{x} and should return a logical value. For example, to use the condition "If the parent's first trait value is negative, do \code{how} (i.e. return \code{TRUE}):

    \code{when = function(x) return(x < 0)} 

\code{how} must be a function that inputs at least the parent trait value(s) as \code{x} and should return the value associated to \code{what} (i.e. a \code{numeric} value for \code{"waiting"} or a \code{logical} value for \code{"speciating"}. For example, to always allow speciation if \code{when} is \code{TRUE}:

     \code{how = function(x) return(TRUE)}

NOTE: modifiers only affect the birth death process after the first speciation event.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
