% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.modifiers.R
\name{make.modifiers}
\alias{make.modifiers}
\title{make.modifiers}
\usage{
make.modifiers(
  branch.length,
  speciation,
  condition,
  modify,
  add,
  test = TRUE,
  ...
)
}
\arguments{
\item{branch.length}{A function for the waiting time generating branch length (see details).}

\item{speciation}{A function for triggering the speciation events (see details).}

\item{condition}{A function giving the condition on which to modify the output of \code{branch.length} or \code{speciation} (see details). If missing the condition is always met (set to \code{condition = function(x, ...) TRUE})}

\item{modify}{A function giving the rule of how to modify the output of \code{branch.length} or \code{speciation} (see details). If missing no modification is used (set to \code{modify = function(x, ...) x}).}

\item{add}{Whether to add this modifier to a \code{"dads"} \code{"modifier"} object.}

\item{test}{Logical whether to test if the modifiers object will work (default is TRUE),}

\item{...}{Optional, any extra argument to be passed to the modifier function.}
}
\description{
Making modifiers objects for dads based on an ancestor's (parent) trait.
}
\details{

}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
