% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.traits.R
\name{make.traits}
\alias{make.traits}
\title{make.traits}
\usage{
make.traits()
}
\arguments{
\item{n}{The number of traits (default is 1)}

\item{start}{The starting values for each traits (default is 0)}

\item{process}{The trait process(es) (default is process.BM)}

\item{process.args}{An optional named list of optional arguments for the trait process}

\item{names}{The optional names of the traits (generated by default from the process)}

\item{test}{Logical whether to test if the traits object will work (default is TRUE)}
}
\description{
Making traits objects for dads
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
