% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.traits.R
\name{make.traits}
\alias{make.traits}
\title{make.traits}
\usage{
make.traits(
  process = BM.process,
  n,
  start,
  process.args,
  names,
  add,
  test = TRUE
)
}
\arguments{
\item{process}{The trait process(es) (default is \code{\link{BM.process}}).}

\item{n}{Optional, the number of traits per process (default is \code{1}).}

\item{start}{Optional, the starting values for each traits (default is {0}).}

\item{process.args}{Optional, a named list of optional arguments for the trait process.}

\item{names}{Optional, the name(s) of the trait(s) (if missing the name of the process is used to generate the names).}

\item{add}{Optional, another previous \code{"dads"} traits object to which to add the trait.}

\item{test}{Logical, whether to test if the traits object will work with \code{\link{dads}} (\code{TRUE} - default).}
}
\description{
Making traits objects for dads
}
\examples{
## A simple Brownian motion trait (default)
make.traits()

## Two independent Brownian motion traits
make.traits(n = 2)

## Two different traits with different process
## (Brownian motion and Ornstein-Uhlenbeck)
make.traits(process = list(BM.process, step.OU))

## A multidimensional Brownian motion trait with correlation
## and different starting points
(my_traits <- make.traits(n = 4, start = c(0, 1, 2, 3),
                          process.args = list(Sigma = diag(1))))

## Adding a Ornstein-Uhlenbeck trait to the previous trait object
make.traits(process = step.OU, names = "OU_trait",
            add = my_traits)

}
\seealso{

}
\author{
Thomas Guillerme
}
