% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parent.traits}
\alias{parent.traits}
\title{Get parent traits}
\usage{
parent.traits(trait.values, lineage, current = TRUE)
}
\arguments{
\item{trait.values}{The internal table of trait values}

\item{lineage}{The internal lineage data list}

\item{current}{Whether to consider only the current lineage (\code{TRUE} - default) or all the living lineages (\code{FALSE}).}
}
\description{
An internal utility function for \code{modifiers}, \code{traits} or \code{events} to access the value(s) of the parent traits in the \code{dads} algorithm
}
\details{
This function is designed to be used internally in \code{dads} to help \code{modifiers}, \code{traits} or \code{events} objects to access the parent traits of the lineages simulated through the internal birth death algorithm.
}
\seealso{
\code{\link{dads}}
}
\author{
Thomas Guillerme
}
